# fix coreml ANE optimized encoder