Transpose the result back to the format that's accepted by the decoder. 
I tested with tiny, small, and base models, and ran `./tests/run-tests.sh`, result all looks good. @ggerganov  I am not sure why your previous attempt didn't work,  can you double-check?

Performance-wise, this is my result on M3 pro with a 30min audio and the base model(I used a longer audio to get a better average encode time per segment so you can ignore the inital coreml model load overhead). The encode time is ~2x faster than metal.
With ANE optimized model:
```
whisper_print_timings:     load time =    93.50 ms
whisper_print_timings:     fallbacks =   0 p /   0 h
whisper_print_timings:      mel time =   678.59 ms
whisper_print_timings:   sample time =  6985.69 ms / 36729 runs (    0.19 ms per run)
whisper_print_timings:   encode time =  1985.48 ms /    67 runs (   29.63 ms per run)
whisper_print_timings:   decode time =   210.62 ms /    86 runs (    2.45 ms per run)
whisper_print_timings:   batchd time = 29357.13 ms / 36314 runs (    0.81 ms per run)
whisper_print_timings:   prompt time =   857.84 ms / 14712 runs (    0.06 ms per run)
whisper_print_timings:    total time = 42247.07 ms
```

With vanilla openai whisper model:
```
whisper_print_timings:     load time =    97.97 ms
whisper_print_timings:     fallbacks =   0 p /   0 h
whisper_print_timings:      mel time =   671.39 ms
whisper_print_timings:   sample time =  7039.62 ms / 36803 runs (    0.19 ms per run)
whisper_print_timings:   encode time =  2792.66 ms /    67 runs (   41.68 ms per run)
whisper_print_timings:   decode time =   202.56 ms /    84 runs (    2.41 ms per run)
whisper_print_timings:   batchd time = 29273.17 ms / 36390 runs (    0.80 ms per run)
whisper_print_timings:   prompt time =   845.34 ms / 14712 runs (    0.06 ms per run)
whisper_print_timings:    total time = 42520.91 ms
```
Metal:
```
whisper_print_timings:     load time =   103.68 ms
whisper_print_timings:     fallbacks =   0 p /   0 h
whisper_print_timings:      mel time =   678.44 ms
whisper_print_timings:   sample time =  6981.09 ms / 36940 runs (    0.19 ms per run)
whisper_print_timings:   encode time =  3958.86 ms /    66 runs (   59.98 ms per run)
whisper_print_timings:   decode time =   164.65 ms /    67 runs (    2.46 ms per run)
whisper_print_timings:   batchd time = 29736.41 ms / 36546 runs (    0.81 ms per run)
whisper_print_timings:   prompt time =   844.35 ms / 14712 runs (    0.06 ms per run)
whisper_print_timings:    total time = 42517.96 ms
```