@ggerganov 2024-01-03
Thanks for looking into this - I will recheck the results now

For reference, here is the discussion back then: https://github.com/ggerganov/whisper.cpp/discussions/548#discussioncomment-5622733
Somehow the results were corrupted, but either I was doing something wrong or something else got fixed along the way