@ggerganov 2024-01-04
Indeed, the ANE-optimized Core ML models work correctly and are faster than the original models. Here are the results that I get on M2 Ultra (only the "Enc." column is relevant for this change):

### master + Core ML ANE
|    CPU |     OS |           Config |         Model |  Th |    Enc. |    Dec. |    Bch5 |      PP |  Commit |
|    --- |    --- |              --- |           --- | --- |     --- |     --- |     --- |     --- |     --- |
| <todo> | <todo> |  NEON BLAS COREML METAL |          tiny |   4 |   29.37 |    1.37 |    0.51 |    0.01 | eaac005 |
| <todo> | <todo> |  NEON BLAS COREML METAL |          base |   4 |   43.33 |    1.98 |    0.78 |    0.02 | eaac005 |
| <todo> | <todo> |  NEON BLAS COREML METAL |         small |   4 |  139.86 |    3.88 |    1.71 |    0.05 | eaac005 |
| <todo> | <todo> |  NEON BLAS COREML METAL |        medium |   4 |  658.38 |    8.09 |    3.93 |    0.13 | eaac005 |
| <todo> | <todo> |  NEON BLAS COREML METAL |      large-v2 |   4 | 1686.48 |   11.68 |    6.06 |    0.23 | eaac005 |

### PR + Core ML ANE
|    CPU |     OS |           Config |         Model |  Th |    Enc. |    Dec. |    Bch5 |      PP |  Commit |
|    --- |    --- |              --- |           --- | --- |     --- |     --- |     --- |     --- |     --- |
| <todo> | <todo> |  NEON BLAS COREML METAL |          tiny |   4 |   23.17 |    1.37 |    0.51 |    0.01 | eaac005 |
| <todo> | <todo> |  NEON BLAS COREML METAL |          base |   4 |   30.34 |    1.97 |    0.77 |    0.02 | eaac005 |
| <todo> | <todo> |  NEON BLAS COREML METAL |         small |   4 |   99.59 |    3.96 |    1.75 |    0.05 | eaac005 |
| <todo> | <todo> |  NEON BLAS COREML METAL |        medium |   4 |  307.01 |    8.06 |    3.93 |    0.13 | eaac005 |
| <todo> | <todo> |  NEON BLAS COREML METAL |      large-v2 |   4 |  526.80 |   11.68 |    6.03 |    0.23 | eaac005 |

Notice however that the ANE-optimized Core ML models are not suitable to run on the GPU (i.e. `config.computeUnits = MLComputeUnitsCPUAndGPU;`):

master + Core ML GPU
|    CPU |     OS |           Config |         Model |  Th |    Enc. |    Dec. |    Bch5 |      PP |  Commit |
|    --- |    --- |              --- |           --- | --- |     --- |     --- |     --- |     --- |     --- |
| <todo> | <todo> |  NEON BLAS COREML METAL |          tiny |   4 |    6.93 |    1.37 |    0.51 |    0.01 | eaac005 |
| <todo> | <todo> |  NEON BLAS COREML METAL |          base |   4 |   12.80 |    1.97 |    0.77 |    0.02 | eaac005 |
| <todo> | <todo> |  NEON BLAS COREML METAL |         small |   4 |   31.70 |    3.95 |    1.71 |    0.05 | eaac005 |
| <todo> | <todo> |  NEON BLAS COREML METAL |        medium |   4 |   93.59 |    8.09 |    3.93 |    0.13 | eaac005 |
| <todo> | <todo> |  NEON BLAS COREML METAL |      large-v2 |   4 |  179.49 |   11.69 |    5.93 |    0.23 | eaac005 |

PR + Core ML GPU
|    CPU |     OS |           Config |         Model |  Th |    Enc. |    Dec. |    Bch5 |      PP |  Commit |
|    --- |    --- |              --- |           --- | --- |     --- |     --- |     --- |     --- |     --- |
| <todo> | <todo> |  NEON BLAS COREML METAL |          tiny |   4 |    9.21 |    1.37 |    0.51 |    0.01 | eaac005 |
| <todo> | <todo> |  NEON BLAS COREML METAL |          base |   4 |   18.46 |    1.97 |    0.78 |    0.02 | eaac005 |
| <todo> | <todo> |  NEON BLAS COREML METAL |         small |   4 |   59.12 |    3.97 |    1.72 |    0.05 | eaac005 |
| <todo> | <todo> |  NEON BLAS COREML METAL |        medium |   4 |  168.24 |    8.08 |    3.87 |    0.13 | eaac005 |
| <todo> | <todo> |  NEON BLAS COREML METAL |      large-v2 |   4 |  303.07 |   11.72 |    5.95 |    0.23 | eaac005 |


For reference, here are the results for running the entire computation on the GPU with Metal (i.e. no Core ML):

Full Metal (no Core ML)
|    CPU |     OS |           Config |         Model |  Th |    Enc. |    Dec. |    Bch5 |      PP |  Commit |
|    --- |    --- |              --- |           --- | --- |     --- |     --- |     --- |     --- |     --- |
| <todo> | <todo> |  NEON BLAS METAL |          tiny |   4 |   10.74 |    1.37 |    0.51 |    0.01 | eaac005 |
| <todo> | <todo> |  NEON BLAS METAL |          base |   4 |   19.06 |    1.98 |    0.78 |    0.02 | eaac005 |
| <todo> | <todo> |  NEON BLAS METAL |         small |   4 |   53.24 |    3.87 |    1.71 |    0.05 | eaac005 |
| <todo> | <todo> |  NEON BLAS METAL |        medium |   4 |  143.67 |    8.12 |    3.94 |    0.13 | eaac005 |
| <todo> | <todo> |  NEON BLAS METAL |      large-v2 |   4 |  253.30 |   11.67 |    6.06 |    0.23 | eaac005 |