@Josscii 2024-01-05
2024-01-05 09:24:22.023112+0800 [4575:1504107] Error: Transpose unit is not supported.
2024-01-05 09:24:22.023235+0800 [4575:1504107] Error: Transpose unit is not supported.
2024-01-05 09:24:22.023298+0800 [4575:1504107] Error: Transpose unit is not supported.
2024-01-05 09:24:22.034869+0800 [4575:1504107] Error: Transpose unit is not supported.
2024-01-05 09:24:22.035000+0800 [4575:1504107] Error: Transpose unit is not supported.
2024-01-05 09:24:22.035063+0800 [4575:1504107] Error: Transpose unit is not supported.
2024-01-05 09:24:27.377632+0800 [4575:1504107] [espresso] [Espresso::handle_ex_plan] exception=at at /ggml-base-encoder.mlmodelc/model.mil:14:12: In 'ios16.conv' operations, tensors parameter x[0], parameter weight[0], parameter bias[0], and output at index 0 must have the same data type.
2024-01-05 09:24:27.377830+0800 [4575:1504107] [coreml] Error plan build: -1.

---

with this update, I generated new coreml encoder and run on iPhone XR with iOS 16, it output above errors