@ggerganov 2024-01-05
Hm, interesting. I actually didn't test if the ANE models work on iOS. Maybe this is the problem that we observed in the past, and now there is an error actually being reported